Move net::{data_loader,image_cache_task} unit tests into the unit_tests crate.
