+++ components/net/data_loader.rs
@@ -21,7 +21,7 @@ pub fn factory(load_data: LoadData, _classifier: Arc<MIMEClassifier>) {
     load(load_data)
 }
 
-fn load(load_data: LoadData) {
+pub fn load(load_data: LoadData) {
     let start_chan = load_data.consumer;
     let url = load_data.url;
     assert!(&*url.scheme == "data");
@@ -80,72 +80,4 @@ fn load(load_data: LoadData) {
         progress_chan.send(Payload(bytes)).unwrap();
         progress_chan.send(Done(Ok(()))).unwrap();
     }
-}
-
-#[cfg(test)]
-fn assert_parse(url:          &'static str,
-                content_type: Option<(String, String)>,
-                charset:      Option<String>,
-                data:         Option<Vec<u8>>) {
-    use std::sync::mpsc::channel;
-    use url::Url;
-
-    let (start_chan, start_port) = channel();
-    load(LoadData::new(Url::parse(url).unwrap(), start_chan));
-
-    let response = start_port.recv().unwrap();
-    assert_eq!(&response.metadata.content_type, &content_type);
-    assert_eq!(&response.metadata.charset,      &charset);
-
-    let progress = response.progress_port.recv().unwrap();
-
-    match data {
-        None => {
-            assert_eq!(progress, Done(Err("invalid data uri".to_string())));
-        }
-        Some(dat) => {
-            assert_eq!(progress, Payload(dat));
-            assert_eq!(response.progress_port.recv().unwrap(), Done(Ok(())));
-        }
-    }
-}
-
-#[test]
-fn empty_invalid() {
-    assert_parse("data:", None, None, None);
-}
-
-#[test]
-fn plain() {
-    assert_parse("data:,hello%20world", None, None, Some(b"hello world".iter().map(|&x| x).collect()));
-}
-
-#[test]
-fn plain_ct() {
-    assert_parse("data:text/plain,hello",
-        Some(("text".to_string(), "plain".to_string())), None, Some(b"hello".iter().map(|&x| x).collect()));
-}
-
-#[test]
-fn plain_charset() {
-    assert_parse("data:text/plain;charset=latin1,hello",
-        Some(("text".to_string(), "plain".to_string())), Some("latin1".to_string()), Some(b"hello".iter().map(|&x| x).collect()));
-}
-
-#[test]
-fn base64() {
-    assert_parse("data:;base64,C62+7w==", None, None, Some(vec!(0x0B, 0xAD, 0xBE, 0xEF)));
-}
-
-#[test]
-fn base64_ct() {
-    assert_parse("data:application/octet-stream;base64,C62+7w==",
-        Some(("application".to_string(), "octet-stream".to_string())), None, Some(vec!(0x0B, 0xAD, 0xBE, 0xEF)));
-}
-
-#[test]
-fn base64_charset() {
-    assert_parse("data:text/plain;charset=koi8-r;base64,8PLl9+XkIO3l5Pfl5A==",
-        Some(("text".to_string(), "plain".to_string())), Some("koi8-r".to_string()),
-        Some(vec!(0xF0, 0xF2, 0xE5, 0xF7, 0xE5, 0xE4, 0x20, 0xED, 0xE5, 0xE4, 0xF7, 0xE5, 0xE4)));
 }
