+++ tests/unit/lib.rs
@@ -2,7 +2,9 @@ extern crate geom;
 extern crate gfx;
 extern crate net;
 extern crate net_traits;
+extern crate profile;
 extern crate util;
+extern crate url;
 
 #[cfg(test)] #[path="gfx/mod.rs"] mod gfx_tests;
 #[cfg(test)] #[path="net/mod.rs"] mod net_tests;
