+++ tests/unit/net/mod.rs
@@ -1,1 +1,3 @@
 mod cookie;
+mod data_loader;
+mod image_cache_task;
